+++ sihl/queue/queue.ml
@@ -1,8 +1,8 @@
-module Service = Queue_service
-module Job = Queue_core.Job
-module JobInstance = Queue_core.JobInstance
-module WorkableJob = Queue_core.WorkableJob
+module Job = Model.Job
+module JobInstance = Model.JobInstance
+module WorkableJob = Model.WorkableJob
+module Sig = Sig
 
-let create_job = Queue_core.Job.create
-let set_max_tries = Queue_core.Job.set_max_tries
-let set_retry_delay = Queue_core.Job.set_retry_delay
+let create_job = Model.Job.create
+let set_max_tries = Model.Job.set_max_tries
+let set_retry_delay = Model.Job.set_retry_delay